Separate spec files for tests by block
